Allow the debugger to write
Thanks to the new MMU architecture, the feature integrate pretty easly.

The instruction cache is updated so writing breakpoints or editing
instruction should works.

Writing to P4 area memory do not update the instruction cache beacause
the module's function should update it since it embed some metadata
probably about some system stuff. No idea if it is possible to run code
on the on-chip ram nor if updating it through the debugger works.